requirements.txt
maltego-trx
# Official Tomba.io Python SDK
tomba-io
